Changed longs and floats to doubles and compensated by using a greater gravity divisor. Fixed typo (diviser/divisor).

+++ src/GravitationalConstants.java
@@ -1,4 +1,4 @@
 public interface GravitationalConstants {
 	
-	long gravitationalConstant = (long)(6.674 * (10^-11));
+	double gravitationalConstant = (6.674 * (10^-11));
 }
